+++ 07-tic-tac-toe-starting-project/src/components/GameBoard.jsx
@@ -1,37 +1,23 @@
-import { useState } from "react";
 
-const intialGameboard = [
-    [null, null, null],
-    [null, null, null],
-    [null, null, null],        
-]
 
-export default function GameBoard() {
-
-    const [gameBoard, setGameBoard] = useState(intialGameboard);
-
-    function clickHandler(rowIndex, cellIndex) {
-        setGameBoard((previousGameboard) => {
-            const newBoard = [...previousGameboard];
-            newBoard[rowIndex][cellIndex] = "X";   
-            return newBoard; 
-        });
-    }    
+export default function GameBoard({onSelectSquare, board}) {
 
     return (
         <ol id="game-board">
-                {gameBoard.map((row, rowIndex) => (
-                    <li key={rowIndex}> 
-                        <ol>
-                            {row.map((cell, cellIndex) => (
-                                <li>
-                                    <button onClick={() => clickHandler(rowIndex, cellIndex)}>{cell}</button>
-                                </li>
-                            ))}
-                        </ol>
-                    </li>  
-                ))}
-
+            {board.map((row, rowIndex) => {
+                return <li key={rowIndex}>
+                    <ol>
+                        {row.map((playerSymbol, colIndex) => {
+                            return <li key={colIndex}>
+                                <button onClick={() => onSelectSquare(rowIndex, colIndex)} disabled={playerSymbol != null}>
+                                    {playerSymbol}
+                                </button>
+                            </li>
+                        })}
+                    </ol>
+                </li>
+            })}
         </ol>
-    );
+    )
 }
+
