--- 07-tic-tac-toe-starting-project/src/components/Gameboard.jsx
@@ -1,23 +1,0 @@
-
-
-export default function GameBoard({onSelectSquare, board}) {
-
-    return (
-        <ol id="game-board">
-            {board.map((row, rowIndex) => {
-                return <li key={rowIndex}>
-                    <ol>
-                        {row.map((playerSymbol, colIndex) => {
-                            return <li key={colIndex}>
-                                <button onClick={() => onSelectSquare(rowIndex, colIndex)} disabled={playerSymbol != null}>
-                                    {playerSymbol}
-                                </button>
-                            </li>
-                        })}
-                    </ol>
-                </li>
-            })}
-        </ol>
-    )
-}
-
